refactor: 웨딩 생성을 전략패턴으로 개선
